renamed and moved scene loading logic file
