+++ .env
@@ -1,6 +1,6 @@
 -LOCAL_SCENE_PATH=../sdk7-scene-template/bin/index.js
 
-REMOTE_SCENE_COORDS=0,0
--REMOTE_SCENE_COORDS=100,100
+-REMOTE_SCENE_COORDS=0,0
+REMOTE_SCENE_COORDS=100,100
 -REMOTE_SCENE_COORDS=60,-137
 -REMOTE_SCENE_COORDS=-129,-77
