+++ src/service.ts
@@ -1,6 +1,6 @@
 import { Lifecycle } from '@well-known-components/interfaces'
 import { BaseComponents } from './types'
-import { loadOrReload } from './controllers/sceneLoadController'
+import { loadOrReload } from './logic/sceneLoader'
 
 // this function wires the business logic (adapters & controllers) with the components (ports)
 export async function main(program: Lifecycle.EntryPointParameters<BaseComponents>) {
